==> 01-get_coverage.ipynb <==
import numpy as np
from coverage import gen_sats, gen_times, camera_model, forecast_fovs, calculate_revisits, revisit_map
from datetime import datetime, timezone, timedelta
import dataclasses
from shapely.geometry import Point
import geopandas as gpd
import matplotlib.pyplot as plt
import pandas as pd
import branca
import folium

from skyfield.framelib import itrs

# from landsat import Instrument, Platform, Scene
from vanzyl1 import Instrument, Platform, Scene

! mkdir tmp

start_dt = datetime.fromisoformat(Scene.start_utc)
num_days = 50

tles = gen_sats(
    sat_nos=[Platform.norad_id] # How to best handle multiple platforms? (TLE vs. SPG4 model too)
)

inst = camera_model(
    name=Instrument.name, 
    fl=Instrument.focal_length_mm, 
    pitch=Instrument.pitch_um*1e-3, 
    h_pix=Instrument.rows, 
    v_pix=Instrument.cols, 
)

times = gen_times(
    start_yr=start_dt.year,
    start_mo=start_dt.month, 
    start_day=start_dt.day, 
    days=num_days, 
    step_min=Instrument.img_period)

xcell_size = ycell_size = .1

## Batch FOV generation over N satellites - TODO: build multiple sats into config/ main script
gdfs = []
for tle in tles:
    sat = tle[0]

    fov_df = forecast_fovs(sat, times, inst)
    xyz_dist_rates = sat.at(times).frame_xyz_and_velocity(itrs)
    fov_df['x_pos'], fov_df['y_pos'], fov_df['z_pos'] = xyz_dist_rates[0].km
    fov_df['x_vel'], fov_df['y_vel'], fov_df['z_vel'] = xyz_dist_rates[1].km_per_s
    gdfs.append(fov_df)

fov_df = gpd.GeoDataFrame(pd.concat(gdfs, ignore_index=True), crs="EPSG:4326")

fov_df["frac_days"] = fov_df.datetime.dt.dayofweek + fov_df.datetime.dt.hour/24 + fov_df.datetime.dt.minute/(24*60) + fov_df.datetime.dt.second/(24*60*60)
fov_df['time_gap'] = fov_df['frac_days'] - fov_df['frac_days'].shift(1)
fov_df[["datetime", "satellite", "id", "time", "geometry"]].dropna()

# %%timeit
# get_inst_fov(sat, times[0], inst)

## Set ascending/ descending direction
conditions = [
    fov_df['z_vel'].lt(0),
    fov_df['z_vel'].gt(0)
]

choices = ['dsc','asc']
fov_df['asc_dsc'] = np.select(conditions, choices, default='undefined')

## Drop ascending pass FOVs
fov_df.loc[fov_df.asc_dsc=="asc", "geometry"] = None

## Select AOI from gpd naturalearth dataset (filter by .name for country, .continent for continent)
world = gpd.read_file(gpd.datasets.get_path('naturalearth_lowres'))

# Drop Antarctica
world = world[world.continent != "Antarctica"]

## Drop any FOVs not over land area
join = fov_df.sjoin(world, how="left").dropna()
fov_df.loc[join.index, 'mode'] = "SCIENCE"
fov_df.loc[~fov_df.index.isin(join.index), 'mode'] = "STANDBY"
fov_df.loc[fov_df['mode']=="STANDBY", "geometry"] = None

## Plotting FOVs

## Create cmap for unique satellites and create color column
## TODO: Move to plotting function?
sat_ids = list(fov_df["id"].unique()).sort()
cmap = branca.colormap.StepColormap(['red', 'blue'], sat_ids, vmin=139084, vmax = 149260)
fov_df['color'] = fov_df['id'].apply(cmap)

## Make a folium map
# m1 = fov_df[fov_df['mode']=="SCIENCE"].drop('datetime', axis=1).explore(color="color", style_kwds={'fillOpacity':0.2}, tooltip=["satellite", "time"])

## View
# m1

## Coverage data analysis for single satellite/ batch of satellites

## Set AOI
# aoi =  world[world.name == "Brazil"].geometry
# aoi =  world[world.continent == "North America"].geometry
# aoi = world[world.name == "United States of America"].geometry # Includes Alaska...

## Or read in aoi from .geojson
aoi = gpd.read_file('./aois/eastern_us.geojson').geometry # ...so use AOI for subsection of US

## Filter fov_df by aoi
xmin, ymin, xmax, ymax= aoi.total_bounds
revisit_df = fov_df.cx[xmin: xmax, ymin: ymax]

## Create revisit map on regular grid
grid, grid_shape = calculate_revisits(revisit_df, aoi, grid_x=xcell_size, grid_y=ycell_size)
grid.n_visits.fillna(0).describe()
m2 = revisit_map(grid, grid_shape, grid_x=xcell_size, grid_y=ycell_size)

# # Add WRS2 - Only use with AOI applied!
# wrs2 = gpd.read_file('./WRS2_descending_0/WRS2_descending.shp')
# wrs2 = wrs2.cx[xmin: xmax, ymin: ymax]
# folium.GeoJson(data=wrs2["geometry"], overlay=False).add_to(m)

m2#.save("./tmp/{}_revisit_map.html".format(satname))

# TODO: Fix Landsat image timing to match WRS-2 grid, as described here: https://landsat.gsfc.nasa.gov/about/the-worldwide-reference-system/

# TODO: Fix saves to GeoJSON (done using patch below)
import fiona

# grid.to_file('./tmp/all_revisits.geojson')

save_df = fov_df.loc[:,['satellite', 'geometry', 'time']].dropna()
save_df['satellite'] = save_df["satellite"].astype(str)
save_df['time'] = save_df["time"].astype(str)

## Save to geojson based on sat name
for satname in save_df.satellite.unique():
    with fiona.Env(OSR_WKT_FORMAT="WKT2_2018"):
        save_df[save_df.satellite==satname].to_file("./tmp/{}_fovs.geojson".format(satname.replace(" ", "_")), engine="pyogrio")

# m1.save("./tmp/{}_{}dfov_map.html".format(satname, num_days))

m2.save("./tmp/{}_{}drevisit_map.html".format(satname, num_days))

sat.model.inclo*(180/np.pi)

sat.model.altp*6371
