==> 03-underflight.ipynb <==
http://www.gisandbeers.com/GeoBazar/Libros/Teledeteccion/Manual-Landsat-9-Handbook.pdf

https://www.usgs.gov/centers/eros/news/underfly-positions-landsat-9-below-landsat-8-simultaneous-scenes

| Landsat 8 Cycle Day | Date DOY | Landsat 8 Paths | Percent Path Overlap |
| --- | --- | --- | --- |
|14 Nov 12, 2021 | 316 | 105,121,137,153,169,185,201,217,233,16,32,48,64,80,96 | ~10 percent western side |
|15 Nov 13, 2021 | 317 | 112,128,144,160,176,192,208,224, 7,23,39,55,71,87 | ~50 percent western side |
|16 Nov 14, 2021 | 318 | 103,119,135,151,167,183,199,215,231,14,30,45,62,78,94 | 100 percent overlap |
|1 Nov 15, 2021 | 319 | 110,126,142,158,174,190,206,222,5,21,37,53,69,85 | ~50 percent eastern side |
|2 Nov 16, 2021 | 320 | 101,117,133,149,165,181,197,213,229,12,28,44,60,76,92 | ~10 percent eastern side |

# f = open("./historical_tles/sat39084.txt", "r")

# tles = []
# for x in f:
#   if x[0] == "1":
#         # print(x) 
#         tle1 = x
#   if x[0] == "2":
#         # print(x) 
#         tle2 = x
#         tle = tle1 + tle2
#         tles.append(tle)
