==> 99-tsgeojson.ipynb <==
# https://python-visualization.github.io/folium/plugins.html

from folium.plugins import TimestampedGeoJson, TimestampedWmsTileLayers
from folium import WmsTileLayer
import folium

# data = TimestampedGeoJson({

#     'type': 'FeatureCollection',

#     "features": [
#         { "type": "Feature", "properties": { "satellite": "LANDSAT 8", "id": 139084, "time": "2022-06-10 00:41:00 UTC", "lonspan": 10.430223652592815, "color": "#ff0000ff" }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 174.319268082861441, 72.98510219076141 ], [ 169.272977210535373, 73.85491251506042 ], [ 163.889044430268626, 70.487827821754209 ], [ 168.342659980479169, 69.758984481845999 ], [ 174.319268082861441, 72.98510219076141 ] ] ] } },
#         { "type": "Feature", "properties": { "satellite": "LANDSAT 8", "id": 139084, "time": "2022-06-10 00:42:00 UTC", "lonspan": 8.3866104906389296, "color": "#ff0000ff" }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 168.4418036910306, 69.822680676311563 ], [ 163.993669159979504, 70.567607569742677 ], [ 160.05519320039167, 67.102042556529469 ], [ 163.994118914113329, 66.463034865192469 ], [ 168.4418036910306, 69.822680676311563 ] ] ] } },
#         { "type": "Feature", "properties": { "satellite": "LANDSAT 8", "id": 139084, "time": "2022-06-10 00:43:00 UTC", "lonspan": 6.9658246074639862, "color": "#ff0000ff" }, "geometry": { "type": "Polygon", "coordinates": [ [ [ 164.066116278134814, 66.525527163910482 ], [ 160.130114614047471, 67.179982951633164 ], [ 157.100291670670828, 63.654939071543275 ], [ 160.618364926523185, 63.082264998576306 ], [ 164.066116278134814, 66.525527163910482 ] ] ] } }
#     ]

#     }, duration="PT2M")

# f = open('./tmp/fovs.geojson')
# data1 = TimestampedGeoJson(f.read(), 
#     period="PT5M", 
#     duration="PT100M"
# )
# f.close()

# m = folium.Map()
# data1.add_to(m)

# m

f = open('./tmp/LANDSAT_8_fovs.geojson')
data1 = TimestampedGeoJson(f.read(), 
    period="PT5M", 
    duration="PT100M"
)
f.close()

# f = open('./tmp/LANDSAT_9_fovs.geojson')
# data2 = TimestampedGeoJson(f.read(), period="PT5M", duration="PT100M")
# f.close()

m = folium.Map()
data1.add_to(m)
# data2.add_to(m)

m
